api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Cornea backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.0.1,<4.1",
    "fastapi>=0.136.1",
    "google-auth>=2.49.2",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.2",
    "langgraph>=1.1.10",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.27",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]
